Bug Fix
For some reason, making Dugan's JSONs look clean (i.e. adding newlines
and tabs so it looks logical) breaks the parser on the iOS side. This
is an Apple bug, but oh well... This update puts it back to Dugan's
version except for the added nutrition stuff and a couple small changes
for testing.